requirements/core.txt
# Core Dependencies - Essential for basic functionality
# Use pyyaml-include as an alternative to PyYAML for better compatibility
pyyaml-include>=1.3.1
aiohttp>=3.9.1
asyncio>=3.4.3
rich>=13.7.0
loguru>=0.7.2
tqdm>=4.66.1
requests>=2.31.0
colorama>=0.4.6
prettytable>=3.9.0
jinja2>=3.1.2

# Build dependencies
wheel>=0.42.0
setuptools>=69.0.2
cython>=3.0.7
